ml/marten/requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
opencv-python>=4.5.3
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
pydantic>=1.8.2
